Improve odd status code check
The script is Python 3 which uses real division (rather than
integer division), so `a / 100 == 2` is just an overly complex
way to write `a == 200`, and requests provides a bunch of
constants for various status codes so might as well use that.

If the intent was to check if the status code is between 200 and
300, I would suggest `status in range(200, 300)` or `200 <=
status < 300`. Request's symbolic constants for these bounds are
respectively `requests.codes.ok` and
`requests.codes.multiple_choices`.